move certif scripts into dedicated folder
