+++ api/scripts/certification/cancel-certifications-and-publish-sessions.js
