+++ api/scripts/certification/generate-certificate-verification-code.js
