+++ api/tests/acceptance/scripts/import-certifications-from-csv_test.js
@@ -1,5 +1,5 @@
 const { expect, nock } = require('../../test-helper');
-const script = require('../../../scripts/import-certifications-from-csv');
+const script = require('../../../scripts/certification/import-certifications-from-csv');
 
 describe('Acceptance | Scripts | import-certifications-from-csv.js', function () {
   describe('#assertFileValidity', function () {
