+++ api/tests/acceptance/scripts/update-certification-infos_test.js
@@ -2,7 +2,7 @@ const { expect, databaseBuilder, knex, sinon } = require('../../test-helper');
 const { writeFile, rm } = require('fs/promises');
 const values = require('lodash/values');
 const logger = require('../../../lib/infrastructure/logger');
-const { updateCertificationInfos, headers } = require('../../../scripts/update-certification-infos');
+const { updateCertificationInfos, headers } = require('../../../scripts/certification/update-certification-infos');
 const dataFile = `${__dirname}/data.csv`;
 const sessionIdsFile = `${__dirname}/sessionIds.csv`;
