+++ api/tests/unit/scripts/get-results-certifications_test.js
@@ -1,5 +1,5 @@
 const { expect } = require('chai');
-const getResultsCertifications = require('../../../../api/scripts/get-results-certifications');
+const getResultsCertifications = require('../../../scripts/certification/get-results-certifications');
 
 describe('Unit | Scripts | get-results-certifications.js', function () {
   const HEADERS = [
